reorg files and add AddCommsToTaskList
