+++ example/sparse_advection/sparse_advection_driver.cpp
@@ -16,13 +16,13 @@
 #include <vector>
 
 // Local Includes
+#include "amr_criteria/refinement_package.hpp"
 #include "bvals/cc/bvals_cc_in_one.hpp"
 #include "interface/metadata.hpp"
 #include "interface/update.hpp"
 #include "mesh/meshblock_pack.hpp"
-#include "mesh/refinement_in_one.hpp"
 #include "parthenon/driver.hpp"
-#include "refinement/refinement.hpp"
+#include "prolong_restrict/prolong_restrict.hpp"
 #include "sparse_advection_driver.hpp"
 #include "sparse_advection_package.hpp"
 
@@ -126,18 +126,8 @@ TaskCollection SparseAdvectionDriver::MakeTaskCollection(BlockList_t &blocks,
                              mdudt.get(), beta * dt, mc1.get());
 
     // do boundary exchange
-    auto send =
-        tl.AddTask(update, parthenon::cell_centered_bvars::SendBoundaryBuffers, mc1);
-    auto recv = tl.AddTask(update | start_bound,
-                           parthenon::cell_centered_bvars::ReceiveBoundaryBuffers, mc1);
-
-    auto set = tl.AddTask(recv, parthenon::cell_centered_bvars::SetBoundaries, mc1);
-
-    auto restrict = set;
-    if (pmesh->multilevel) {
-      restrict =
-          tl.AddTask(set, parthenon::cell_centered_bvars::RestrictGhostHalos, mc1, false);
-    }
+    auto restrict = parthenon::cell_centered_bvars::AddBoundaryExchangeTasks(
+        update, tl, mc1, pmesh->multilevel);
 
     // if this is the last stage, check if we can deallocate any sparse variables
     if (stage == integrator->nstages) {
