+++ src/amr_criteria/amr_criteria.hpp
@@ -1,5 +1,5 @@
 //========================================================================================
-// (C) (or copyright) 2020. Triad National Security, LLC. All rights reserved.
+// (C) (or copyright) 2020-2021. Triad National Security, LLC. All rights reserved.
 //
 // This program was produced under U.S. Government contract 89233218CNA000001 for Los
 // Alamos National Laboratory (LANL), which is operated by Triad National Security, LLC
@@ -10,39 +10,35 @@
 // license in this material to reproduce, prepare derivative works, distribute copies to
 // the public, perform publicly and display publicly, and to permit others to do so.
 //========================================================================================
-
-#ifndef REFINEMENT_REFINEMENT_HPP_
-#define REFINEMENT_REFINEMENT_HPP_
+#ifndef AMR_CRITERIA_AMR_CRITERIA_HPP_
+#define AMR_CRITERIA_AMR_CRITERIA_HPP_
 
 #include <memory>
 #include <string>
 
 #include "defs.hpp"
-#include "parthenon_arrays.hpp"
+#include "interface/meshblock_data.hpp"
 
 namespace parthenon {
 
 class ParameterInput;
-class MeshBlock;
-template <typename T>
-class MeshBlockData;
-template <typename T>
-class MeshData;
-class StateDescriptor;
 
-namespace Refinement {
+struct AMRCriteria {
+  AMRCriteria() = default;
+  virtual ~AMRCriteria() {}
+  virtual AmrTag operator()(const MeshBlockData<Real> *rc) const = 0;
+  std::string field;
+  Real refine_criteria, derefine_criteria;
+  int max_level;
+  static std::shared_ptr<AMRCriteria>
+  MakeAMRCriteria(std::string &criteria, ParameterInput *pin, std::string &block_name);
+};
 
-std::shared_ptr<StateDescriptor> Initialize(ParameterInput *pin);
-template <typename T>
-TaskStatus Tag(T *rc);
-
-AmrTag CheckAllRefinement(MeshBlockData<Real> *rc);
-
-AmrTag FirstDerivative(MeshBlock *pmb, const ParArrayND<Real> &q,
-                       const Real refine_criteria, const Real derefine_criteria);
-
-} // namespace Refinement
+struct AMRFirstDerivative : public AMRCriteria {
+  AMRFirstDerivative(ParameterInput *pin, std::string &block_name);
+  AmrTag operator()(const MeshBlockData<Real> *rc) const override;
+};
 
 } // namespace parthenon
 
-#endif // REFINEMENT_REFINEMENT_HPP_
+#endif // AMR_CRITERIA_AMR_CRITERIA_HPP_
